Move NonTrade to its own module
