+++ nd2k/formatter.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from .queries import is_a_purchase, is_sending_funds
-from .types import Transaction, CSV, Trade, Swap, NonTrade, KoinlyTag, Exchange
+from .types import Transaction, CSV, Trade, Swap, KoinlyTag, Exchange
+from .nontrade import NonTrade
 
 
 universal_csv_headers = [
